Move Base Button component inside Base folder and update refs
