+++ ui/components/Button/base/base.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Button from '../';
+import Button from './';
 import mdx from '../docs.stories.mdx';
 
 export default {
